sysdumpcollector: change max_workers value for ThreadPool executors
A value of `multiprocessing.cpu_count() - 1` may potentially be 0, which
is a problem.

As the tasks performed by sysdumpcollector are not typically CPU bound,
increase the minimum number of workers and pick the default value that
is used in Python 3.8 (`min(32, os.cpu_count() + 4)`) as this also
avoids creating too many workers on many-core machines.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>