sysdump: log some data collection errors are warning
If collecting the nodes or pods overview fails, it likely indicates a
major connectivity problem with the cluster or something of the sort.
However, if log collection for hubble, for instance, cannot be
collected it might just mean that hubble is not deployed in this cluster
and logging a warning is sufficient in this case.

Signed-off-by: Robin Hahling <robin.hahling@gw-computing.net>