Remove --since default
Fallback to collecting all logs. The rationale is that we will rely on
the user to set any limits if they are interested in a smaller sysdump
collection.

In practice, we've seen that the default 30m was not a good value. This
was because a 30m is very unlikely to give enough context when
troubleshooting. Ofttimes, when we ask for a sysdump, we end up asking
the customer to do it twice, once for the initial, then another when we
realize the initial doesn't have any useful context.

Fixes: https://github.com/cilium/cilium-sysdump/issues/92

Signed-off-by: Chris Tarazi <chris@isovalent.com>